remove debug log
