+++ server/src/main/java/org/apache/iotdb/db/query/reader/universal/PriorityMergeReader.java
@@ -20,7 +20,6 @@ package org.apache.iotdb.db.query.reader.universal;
 
 import org.apache.iotdb.tsfile.read.reader.IPointReader;
 import org.apache.iotdb.tsfile.read.TimeValuePair;
-
 import java.io.IOException;
 import java.util.List;
 import java.util.PriorityQueue;
