+++ server/src/test/java/org/apache/iotdb/db/integration/IoTDBAggregationIT.java
@@ -19,15 +19,15 @@
 
 package org.apache.iotdb.db.integration;
 
-import static org.apache.iotdb.db.integration.TestConstant.avg;
-import static org.apache.iotdb.db.integration.TestConstant.count;
-import static org.apache.iotdb.db.integration.TestConstant.first_value;
-import static org.apache.iotdb.db.integration.TestConstant.last_value;
-import static org.apache.iotdb.db.integration.TestConstant.max_time;
-import static org.apache.iotdb.db.integration.TestConstant.max_value;
-import static org.apache.iotdb.db.integration.TestConstant.min_time;
-import static org.apache.iotdb.db.integration.TestConstant.min_value;
-import static org.apache.iotdb.db.integration.TestConstant.sum;
+import static org.apache.iotdb.db.constant.TestConstant.avg;
+import static org.apache.iotdb.db.constant.TestConstant.count;
+import static org.apache.iotdb.db.constant.TestConstant.first_value;
+import static org.apache.iotdb.db.constant.TestConstant.last_value;
+import static org.apache.iotdb.db.constant.TestConstant.max_time;
+import static org.apache.iotdb.db.constant.TestConstant.max_value;
+import static org.apache.iotdb.db.constant.TestConstant.min_time;
+import static org.apache.iotdb.db.constant.TestConstant.min_value;
+import static org.apache.iotdb.db.constant.TestConstant.sum;
 import static org.junit.Assert.fail;
 
 import java.sql.Connection;
