+++ server/src/test/java/org/apache/iotdb/db/integration/IoTDBLargeDataIT.java
@@ -28,6 +28,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
+import org.apache.iotdb.db.constant.TestConstant;
 import org.apache.iotdb.db.utils.EnvironmentUtils;
 import org.apache.iotdb.jdbc.Config;
 import org.apache.iotdb.tsfile.common.conf.TSFileConfig;
@@ -89,7 +90,7 @@ public class IoTDBLargeDataIT {
         .getConnection(Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
         Statement statement = connection.createStatement()) {
 
-      for (String sql : TestConstant.create_sql) {
+      for (String sql : org.apache.iotdb.db.constant.TestConstant.create_sql) {
         statement.execute(sql);
       }
 
@@ -120,10 +121,10 @@ public class IoTDBLargeDataIT {
             .format("insert into root.vehicle.d0(timestamp,s2) values(%s,%s)", time, time % 22);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s3) values(%s,'%s')", time,
-            TestConstant.stringValue[time % 5]);
+            org.apache.iotdb.db.constant.TestConstant.stringValue[time % 5]);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s4) values(%s, %s)", time,
-            TestConstant.booleanValue[time % 2]);
+            org.apache.iotdb.db.constant.TestConstant.booleanValue[time % 2]);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s5) values(%s, %s)", time, time);
         statement.execute(sql);
@@ -175,7 +176,7 @@ public class IoTDBLargeDataIT {
             .format("insert into root.vehicle.d0(timestamp,s2) values(%s,%s)", time, time + 2);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s3) values(%s,'%s')", time,
-            TestConstant.stringValue[time % 5]);
+            org.apache.iotdb.db.constant.TestConstant.stringValue[time % 5]);
         statement.execute(sql);
       }
 
@@ -193,7 +194,7 @@ public class IoTDBLargeDataIT {
             .format("insert into root.vehicle.d0(timestamp,s3) values(%s,'%s')", time, "goodman");
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s4) values(%s, %s)", time,
-            TestConstant.booleanValue[time % 2]);
+            org.apache.iotdb.db.constant.TestConstant.booleanValue[time % 2]);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s5) values(%s, %s)", time, 9999);
         statement.execute(sql);
@@ -222,13 +223,15 @@ public class IoTDBLargeDataIT {
         int cnt = 0;
         while (resultSet.next()) {
           String ans =
-              resultSet.getString(TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
-                  TestConstant.d0s0)
-                  + "," + resultSet.getString(TestConstant.d0s1) + "," + resultSet
-                  .getString(TestConstant.d0s2) + ","
-                  + resultSet.getString(TestConstant.d0s3) + "," + resultSet.getString(TestConstant.d0s4)
+              resultSet.getString(org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s0)
+                  + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s1) + "," + resultSet
+                  .getString(org.apache.iotdb.db.constant.TestConstant.d0s2) + ","
+                  + resultSet.getString(org.apache.iotdb.db.constant.TestConstant.d0s3) + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s4)
                   + ","
-                  + resultSet.getString(TestConstant.d0s5);
+                  + resultSet.getString(org.apache.iotdb.db.constant.TestConstant.d0s5);
           cnt++;
         }
 
@@ -258,8 +261,8 @@ Statement statement = connection.createStatement()) {
         int cnt = 0;
         while (resultSet.next()) {
           String ans =
-              resultSet.getString(TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
-                  TestConstant.d0s0);
+              resultSet.getString(org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s0);
           // System.out.println("===" + ans);
           cnt++;
         }
@@ -290,8 +293,8 @@ Statement statement = connection.createStatement()) {
       try (ResultSet resultSet = statement.getResultSet();) {
         while (resultSet.next()) {
           String ans =
-              resultSet.getString(TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
-                  TestConstant.d0s0);
+              resultSet.getString(org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s0);
           // System.out.println(ans);
           cnt++;
         }
@@ -319,7 +322,8 @@ Statement statement = connection.createStatement()) {
       try (ResultSet resultSet = statement.getResultSet()) {
         int cnt = 0;
         while (resultSet.next()) {
-          long time = Long.valueOf(resultSet.getString(TestConstant.TIMESTAMP_STR));
+          long time = Long.valueOf(resultSet.getString(
+              org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR));
           String value = resultSet.getString(TestConstant.d0s1);
           if (time > 200900) {
             assertEquals("7777", value);
