+++ server/src/test/java/org/apache/iotdb/db/integration/IoTDBMultiSeriesIT.java
@@ -28,6 +28,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
+import org.apache.iotdb.db.constant.TestConstant;
 import org.apache.iotdb.db.utils.EnvironmentUtils;
 import org.apache.iotdb.jdbc.Config;
 import org.apache.iotdb.tsfile.common.conf.TSFileConfig;
@@ -43,7 +44,7 @@ import org.junit.Test;
  */
 public class IoTDBMultiSeriesIT {
 
-  private static boolean testFlag = TestConstant.testFlag;
+  private static boolean testFlag = org.apache.iotdb.db.constant.TestConstant.testFlag;
   private static TSFileConfig tsFileConfig = TSFileDescriptor.getInstance().getConfig();
   private static int maxNumberOfPointsInPage;
   private static int pageSizeInByte;
@@ -91,7 +92,7 @@ public class IoTDBMultiSeriesIT {
         .getConnection(Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
         Statement statement = connection.createStatement()) {
 
-      for (String sql : TestConstant.create_sql) {
+      for (String sql : org.apache.iotdb.db.constant.TestConstant.create_sql) {
         statement.execute(sql);
       }
 
@@ -137,10 +138,10 @@ public class IoTDBMultiSeriesIT {
             .format("insert into root.vehicle.d0(timestamp,s2) values(%s,%s)", time, time % 22);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s3) values(%s,'%s')", time,
-            TestConstant.stringValue[time % 5]);
+            org.apache.iotdb.db.constant.TestConstant.stringValue[time % 5]);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s4) values(%s, %s)", time,
-            TestConstant.booleanValue[time % 2]);
+            org.apache.iotdb.db.constant.TestConstant.booleanValue[time % 2]);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s5) values(%s, %s)", time, time);
         statement.execute(sql);
@@ -192,7 +193,7 @@ public class IoTDBMultiSeriesIT {
             .format("insert into root.vehicle.d0(timestamp,s2) values(%s,%s)", time, time + 2);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s3) values(%s,'%s')", time,
-            TestConstant.stringValue[time % 5]);
+            org.apache.iotdb.db.constant.TestConstant.stringValue[time % 5]);
         statement.execute(sql);
       }
 
@@ -210,7 +211,7 @@ public class IoTDBMultiSeriesIT {
             .format("insert into root.vehicle.d0(timestamp,s3) values(%s,'%s')", time, "goodman");
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s4) values(%s, %s)", time,
-            TestConstant.booleanValue[time % 2]);
+            org.apache.iotdb.db.constant.TestConstant.booleanValue[time % 2]);
         statement.execute(sql);
         sql = String.format("insert into root.vehicle.d0(timestamp,s5) values(%s, %s)", time, 9999);
         statement.execute(sql);
@@ -237,7 +238,7 @@ public class IoTDBMultiSeriesIT {
         int cnt = 0;
         while (resultSet.next()) {
           String ans =
-              resultSet.getString(TestConstant.TIMESTAMP_STR) + "," + resultSet
+              resultSet.getString(org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR) + "," + resultSet
                   .getString("root.fans.d0.s0")
                   + "," + resultSet.getString("root.fans.d0.s1");
           cnt++;
@@ -265,13 +266,15 @@ public class IoTDBMultiSeriesIT {
         int cnt = 0;
         while (resultSet.next()) {
           String ans =
-              resultSet.getString(TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
-                  TestConstant.d0s0)
-                  + "," + resultSet.getString(TestConstant.d0s1) + "," + resultSet
-                  .getString(TestConstant.d0s2) + ","
-                  + resultSet.getString(TestConstant.d0s3) + "," + resultSet.getString(TestConstant.d0s4)
+              resultSet.getString(org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR) + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s0)
+                  + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s1) + "," + resultSet
+                  .getString(org.apache.iotdb.db.constant.TestConstant.d0s2) + ","
+                  + resultSet.getString(org.apache.iotdb.db.constant.TestConstant.d0s3) + "," + resultSet.getString(
+                  org.apache.iotdb.db.constant.TestConstant.d0s4)
                   + ","
-                  + resultSet.getString(TestConstant.d0s5);
+                  + resultSet.getString(org.apache.iotdb.db.constant.TestConstant.d0s5);
           cnt++;
         }
         assertEquals(23400, cnt);
@@ -299,8 +302,8 @@ public class IoTDBMultiSeriesIT {
         int cnt = 0;
         while (resultSet.next()) {
           String ans =
-              resultSet.getString(TestConstant.TIMESTAMP_STR) + "," + resultSet
-                  .getString(TestConstant.d0s0);
+              resultSet.getString(org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR) + "," + resultSet
+                  .getString(org.apache.iotdb.db.constant.TestConstant.d0s0);
           // System.out.println("===" + ans);
           cnt++;
         }
@@ -330,8 +333,8 @@ public class IoTDBMultiSeriesIT {
         int cnt = 0;
         while (resultSet.next()) {
           String ans =
-              resultSet.getString(TestConstant.TIMESTAMP_STR) + "," + resultSet
-                  .getString(TestConstant.d0s0);
+              resultSet.getString(org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR) + "," + resultSet
+                  .getString(org.apache.iotdb.db.constant.TestConstant.d0s0);
           // System.out.println(ans);
           cnt++;
         }
@@ -360,7 +363,8 @@ public class IoTDBMultiSeriesIT {
       try (ResultSet resultSet = statement.getResultSet()) {
         int cnt = 0;
         while (resultSet.next()) {
-          long time = Long.valueOf(resultSet.getString(TestConstant.TIMESTAMP_STR));
+          long time = Long.valueOf(resultSet.getString(
+              org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR));
           String value = resultSet.getString(TestConstant.d0s1);
           if (time > 200900) {
             assertEquals("7777", value);
