+++ server/src/test/java/org/apache/iotdb/db/integration/IoTDBQuotedPathIT.java
@@ -27,7 +27,7 @@ import org.junit.Test;
 import java.sql.*;
 import java.util.ArrayList;
 
-import static org.apache.iotdb.db.integration.TestConstant.*;
+import static org.apache.iotdb.db.constant.TestConstant.*;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
